# internal/routes/openapi.yaml
openapi: 3.0.3
info:
  title: Go Gin Boilerplate API
  version: 1.0.0
  description: REST endpoints for authentication and RBAC. GraphQL has its own schema.
servers:
  - url: /
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive an access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get the current user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current access token
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/rbac/roles:
    get:
      tags: [Roles]
      summary: List roles
      security: [{ bearerAuth: [] }]
      x-required-permission: role:index
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PerPage' }
        - { $ref: '#/components/parameters/Search' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Roles]
      summary: Create a role
      security: [{ bearerAuth: [] }]
      x-required-permission: role:create
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/rbac/roles/{id}:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    put:
      tags: [Roles]
      summary: Update a role
      security: [{ bearerAuth: [] }]
      x-required-permission: role:update
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Roles]
      summary: Delete a role
      security: [{ bearerAuth: [] }]
      x-required-permission: role:delete
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/rbac/roles/{id}/permissions:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    get:
      tags: [Roles]
      summary: Get permissions assigned to a role
      security: [{ bearerAuth: [] }]
      x-required-permission: role:assign_permission
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Roles]
      summary: Assign permissions to a role
      security: [{ bearerAuth: [] }]
      x-required-permission: role:assign_permission
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignPermissionsInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/rbac/permissions:
    get:
      tags: [Permissions]
      summary: List permissions
      security: [{ bearerAuth: [] }]
      x-required-permission: permission:index
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PerPage' }
        - { $ref: '#/components/parameters/Search' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Permissions]
      summary: Create a permission
      security: [{ bearerAuth: [] }]
      x-required-permission: permission:create
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PermissionInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/rbac/permissions/{id}:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    put:
      tags: [Permissions]
      summary: Update a permission
      security: [{ bearerAuth: [] }]
      x-required-permission: permission:update
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PermissionInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Permissions]
      summary: Delete a permission
      security: [{ bearerAuth: [] }]
      x-required-permission: permission:delete
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 100 }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error:
      description: Invalid request or operation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Unauthorized:
      description: Missing or invalid access token
    Forbidden:
      description: Missing permission
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer }
  schemas:
    Envelope:
      type: object
      required: [status, status_code, request_id, message, data, extra_data, errors]
      properties:
        status: { type: boolean, description: True for a successful response }
        status_code: { type: integer, description: Application status or error code }
        request_id: { type: string, format: uuid }
        message: { type: string, description: Localized human-readable message }
        data: {}
        extra_data: {}
        errors: {}
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
...
    RoleInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, maxLength: 100 }
        slug: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 500 }
    PermissionInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, maxLength: 100 }
        slug: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 500 }
    AssignPermissionsInput:
      type: object
      required: [permission_ids]
      properties:
        permission_ids:
          type: array
          items: { type: integer, minimum: 1 }
...